backend/generate_image/requirements.txt
requests>=2.31.0
python-dotenv>=1.0.1
google-api-python-client>=2.140.0
google-auth>=2.35.0

